IGAPP-522: Rename and move useLoadFromEndpoint to hooks
