+++ native/src/routes/offers/containers/__tests__/OffersContainer.spec.js
@@ -7,14 +7,14 @@ import { OFFERS_ROUTE } from 'api-client/src/routes'
 import OffersContainer from '../OffersContainer'
 import { render } from '@testing-library/react-native'
 import ErrorCodes from '../../../../modules/error/ErrorCodes'
-import { useLoadFromEndpoint } from '../../../../modules/endpoint/loadFromEndpoint'
+import { useLoadFromEndpoint } from '../../../../modules/endpoint/hooks/useLoadFromEndpoint'
 import configureMockStore from 'redux-mock-store'
 import CityModelBuilder from 'api-client/src/testing/CityModelBuilder'
 import { CityModel } from 'api-client'
 
 jest.mock('react-i18next')
 jest.mock('../../../../modules/common/openExternalUrl')
-jest.mock('../../../../modules/endpoint/loadFromEndpoint', () => ({
+jest.mock('../../../../modules/endpoint/hooks/useLoadFromEndpoint', () => ({
   useLoadFromEndpoint: jest.fn()
 }))
